add post and reply urls/views/templates and tests.
