+++ accounts/views.py
@@ -2,8 +2,8 @@ from django.contrib.auth import login
 from django.contrib.auth.forms import UserCreationForm
 from django.shortcuts import render
 from django.shortcuts import redirect
-
 from .forms import SignUpForm
+
 
 def signup(request):
     if request.method == 'POST':
