+++ boards/forms.py
@@ -1,5 +1,7 @@
 from django import forms
 from .models import Topic
+from .models import Post
+
 
 class NewTopicForm(forms.ModelForm):
     message = forms.CharField(
@@ -12,3 +14,9 @@ class NewTopicForm(forms.ModelForm):
     class Meta:
         model = Topic
         fields = ['subject', 'message']
+
+
+class PostForm(forms.ModelForm):
+    class Meta:
+        model = Post
+        fields = ['message', ]
