+++ boards/templates/boards/topics.html
@@ -33,7 +33,7 @@
     <tbody>
       {% for topic in board.topics.all %}
         <tr>
-          <td>{{ topic.subject }}</td>
+          <td><a href="{% url 'topic_posts' board.pk topic.pk %}">{{ topic.subject }}</a></td>
           <td>{{ topic.starter.username }}</td>
           <td>0</td>
           <td>0</td>
