+++ boards/tests/test_view_topic_posts.py
@@ -1,0 +1,26 @@
+from django.contrib.auth.models import User
+from django.test import TestCase
+from django.urls import resolve
+from django.urls import reverse
+
+from ..models import Board
+from ..models import Post
+from ..models import Topic
+from ..views import topic_posts
+
+
+class TopicPostsTests(TestCase):
+    def setUp(self):
+        board = Board.objects.create(name='Django', description='Django board.')
+        user = User.objects.create_user(username='john', email='john@doe.com', password='123')
+        topic = Topic.objects.create(subject='Hello, world', board=board, starter=user)
+        Post.objects.create(message='Lorem ipsum dolor sit amet', topic=topic, created_by=user)
+        url = reverse('topic_posts', kwargs={'pk': board.pk, 'topic_pk': topic.pk})
+        self.response = self.client.get(url)
+
+    def test_status_code(self):
+        self.assertEquals(self.response.status_code, 200)
+
+    def test_view_function(self):
+        view = resolve('/boards/1/topics/1/')
+        self.assertEquals(view.func, topic_posts)
