+++ boards/views.py
@@ -4,35 +4,61 @@ from django.shortcuts import render
 from django.shortcuts import redirect
 from django.shortcuts import get_object_or_404
 from .forms import NewTopicForm
+from .forms import PostForm
 from .models import Board
 from .models import Topic
 from .models import Post
+
 
 def home(request):
     boards = Board.objects.all()
     return render(request, 'boards/home.html', {'boards': boards})
 
+
 def board_topics(request, pk):
     board = get_object_or_404(Board, pk=pk)
     return render(request, 'boards/topics.html', {'board': board})
 
+
 @login_required
 def new_topic(request, pk):
     board = get_object_or_404(Board, pk=pk)
-    user = User.objects.first()  # TODO: get the currently logged in user
     if request.method == 'POST':
         form = NewTopicForm(request.POST)
         if form.is_valid():
             topic = form.save(commit=False)
             topic.board = board
-            topic.starter = user
+            topic.starter = request.user
             topic.save()
             post = Post.objects.create(
                 message=form.cleaned_data.get('message'),
                 topic=topic,
-                created_by=user
+                created_by=request.user
             )
-            return redirect('board_topics', pk=board.pk)  # TODO: redirect to the created topic page
+            return redirect('topic_posts', pk=pk, topic_pk=topic.pk)
     else:
         form = NewTopicForm()
-    return render(request, 'boards/new_topic.html', {'board': board, 'form': form})
+    return render(request,
+                  'boards/new_topic.html',
+                  {'board': board, 'form': form})
+
+
+def topic_posts(request, pk, topic_pk):
+    topic = get_object_or_404(Topic, board__pk=pk, pk=topic_pk)
+    return render(request, 'boards/topic_posts.html', {'topic': topic})
+
+
+@login_required
+def reply_topic(request, pk, topic_pk):
+    topic = get_object_or_404(Topic, board__pk=pk, pk=topic_pk)
+    if request.method == 'POST':
+        form = PostForm(request.POST)
+        if form.is_valid():
+            post = form.save(commit=False)
+            post.topic = topic
+            post.created_by = request.user
+            post.save()
+            return redirect('topic_posts', pk=pk, topic_pk=topic_pk)
+    else:
+        form = PostForm()
+    return render(request, 'boards/reply_topic.html', {'topic': topic, 'form': form})
